pyproject.toml
[tool.poetry]
name = "vistabnet"
version = "0.1.0"
description = ""
authors = ["wwydmanski <wwydmanski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.0"
tqdm = "^4.65.0"
focal-loss-torch = "^0.1.2"
torchvision = "^0.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
